% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullScore.R
\name{fullScore}
\alias{fullScore}
\title{Score Function For Zero-Inflated Poisson (ZIP) Distribution}
\usage{
fullScore(x, currentLambda, p, q)
}
\arguments{
\item{x}{A vector containing values belonging to a ZIP Distribution}

\item{currentLambda}{The value of our Lambda Parameter}

\item{p}{Probability of observing a Structural 0}

\item{q}{Probability of not observing a structural 0}
}
\description{
This is the first derivative of Log-Liklihood function for ZIP Distribution
Use this for Newton Raphson Algorithm
}
\examples{
fullScore()
}
