% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtonRaphson_ZIP.R
\name{newtonRaphson_ZIP}
\alias{newtonRaphson_ZIP}
\title{Newton Raphson Function For Zero-Inflated Poisson (ZIP) Distribution}
\usage{
newtonRaphson_ZIP(xVect, delta = 3)
}
\arguments{
\item{xVect}{A vector containing values belonging to a ZIP Distribution}

\item{delta}{Stabilizing Parameter; Default to 3}
}
\description{
This function uses Newton Raphson to Tune the Lambda Parameter of our ZIP Distribution
}
\examples{
newtonRaphson_ZIP()
}
